refactor(cli): extract telemetry delivery into transport.ts
Split the reliability-critical delivery layer — the event queue, async
flush(), and the exit-time detached-child flushSync() — out of client.ts
into a new transport.ts. This is the code that had the exit-race data-loss
bug fixed in #2105; isolating it keeps that subtle path in one focused,
dependency-light module (config + node builtins only).

client.ts stays the CLI-facing policy layer (opt-out checks, system-metadata
enrichment, first-run notice) and re-exports flush/flushSync, so events.ts,
index.ts, and the cli.ts exit handlers keep importing from ./client.js
unchanged.

Pure code motion — public API identical, no behavior change. Telemetry
tests pass unchanged (client.test.ts still exercises delivery through the
public API).
